add base dao and entity support

+++ app/src/main/java/com/kozak/pw/domain/game/PwGame.kt
@@ -7,5 +7,5 @@ import kotlinx.datetime.LocalDateTime
 data class PwGame(
     val createdAt: LocalDateTime,
     val gameSpeed: GameSpeed,
-    val id: Int = PwConstants.DEFAULT_ITEM_ID.toInt()
+    val id: Long = PwConstants.DEFAULT_ITEM_ID
 )
